+++ packages/firebase_auth/lib/src/firebase_user.dart
@@ -60,7 +60,7 @@ class FirebaseUser extends UserInfo {
   ///   • `ERROR_OPERATION_NOT_ALLOWED` - Indicates that this type of account is not enabled.
   ///   • `ERROR_INVALID_ACTION_CODE` - If the action code in the link is malformed, expired, or has already been used.
   ///       This can only occur when using [EmailAuthProvider.getCredentialWithLink] to obtain the credential.
-  Future<FirebaseUser> linkWithCredential(AuthCredential credential) async {
+  Future<AuthResult> linkWithCredential(AuthCredential credential) async {
     assert(credential != null);
     final Map<String, dynamic> data =
         await FirebaseAuth.channel.invokeMapMethod<String, dynamic>(
@@ -71,8 +71,8 @@ class FirebaseUser extends UserInfo {
         'data': credential._data,
       },
     );
-    final FirebaseUser currentUser = FirebaseUser._(data, _app);
-    return currentUser;
+    final AuthResult result = AuthResult._(data, _app);
+    return result;
   }
 
   /// Initiates email verification for the user.
@@ -192,10 +192,11 @@ class FirebaseUser extends UserInfo {
   ///   • `ERROR_USER_DISABLED` - If the user has been disabled (for example, in the Firebase console)
   ///   • `ERROR_USER_NOT_FOUND` - If the user has been deleted (for example, in the Firebase console)
   ///   • `ERROR_OPERATION_NOT_ALLOWED` - Indicates that Email & Password accounts are not enabled.
-  Future<FirebaseUser> reauthenticateWithCredential(
+  Future<AuthResult> reauthenticateWithCredential(
       AuthCredential credential) async {
     assert(credential != null);
-    await FirebaseAuth.channel.invokeMethod<void>(
+    final Map<String, dynamic> data =
+        await FirebaseAuth.channel.invokeMapMethod<String, dynamic>(
       'reauthenticateWithCredential',
       <String, dynamic>{
         'app': _app.name,
@@ -203,7 +204,7 @@ class FirebaseUser extends UserInfo {
         'data': credential._data,
       },
     );
-    return this;
+    return AuthResult._(data, _app);
   }
 
   /// Detaches the [provider] account from the current user.