+++ packages/firebase_auth/test/firebase_auth_test.dart
@@ -32,6 +32,25 @@ const Map<dynamic, dynamic> kMockIdTokenResultClaims = <dynamic, dynamic>{
 const int kMockIdTokenResultExpirationTimestamp = 123456;
 const int kMockIdTokenResultAuthTimestamp = 1234567;
 const int kMockIdTokenResultIssuedAtTimestamp = 12345678;
+const Map<String, dynamic> kMockAdditionalUserInfo = <String, dynamic>{
+  'isNewUser': false,
+  'username': 'flutterUser',
+  'providerId': 'testProvider',
+  'profile': <String, dynamic>{'foo': 'bar'},
+};
+const Map<String, dynamic> kMockUser = <String, dynamic>{
+  'isAnonymous': true,
+  'isEmailVerified': false,
+  'providerData': <Map<String, String>>[
+    <String, String>{
+      'providerId': kMockProviderId,
+      'uid': kMockUid,
+      'displayName': kMockDisplayName,
+      'photoUrl': kMockPhotoUrl,
+      'email': kMockEmail,
+    },
+  ],
+};
 
 void main() {
   group('$FirebaseAuth', () {
@@ -55,6 +74,8 @@ void main() {
           case "startListeningAuthState":
             return mockHandleId++;
             break;
+          case "currentUser":
+            return kMockUser;
           case "sendLinkToEmail":
           case "sendPasswordResetEmail":
           case "updateEmail":
@@ -70,7 +91,10 @@ void main() {
             return null;
             break;
           default:
-            return mockFirebaseUser();
+            return <String, dynamic>{
+              'user': kMockUser,
+              'additionalUserInfo': kMockAdditionalUserInfo,
+            };
             break;
         }
       });
@@ -88,18 +112,18 @@ void main() {
       expect(userInfo.email, kMockEmail);
     }
 
-    test('currentUser', () async {
-      final FirebaseUser user = await auth.currentUser();
-      verifyUser(user);
+    void verifyAuthResult(AuthResult result) {
+      verifyUser(result.user);
+      final AdditionalUserInfo additionalUserInfo = result.additionalUserInfo;
       expect(
-        log,
-        <Matcher>[
-          isMethodCall('currentUser',
-              arguments: <String, String>{'app': auth.app.name}),
-        ],
-      );
-    });
+          additionalUserInfo.isNewUser, kMockAdditionalUserInfo['isNewUser']);
+      expect(additionalUserInfo.username, kMockAdditionalUserInfo['username']);
+      expect(
+          additionalUserInfo.providerId, kMockAdditionalUserInfo['providerId']);
+      expect(additionalUserInfo.profile, kMockAdditionalUserInfo['profile']);
+    }
 
+<<<<<<< HEAD
     void verifyIdTokenResult(IdTokenResult idTokenResult) {
       expect(idTokenResult.token, equals(kMockIdToken));
       expect(
@@ -126,10 +150,16 @@ void main() {
       verifyIdTokenResult(await user.getIdToken());
       verifyIdTokenResult(await user.getIdToken(refresh: true));
 
+=======
+    test('getIdToken', () async {
+      final FirebaseUser user = await auth.currentUser();
+      expect(await user.getIdToken(), equals(kMockIdToken));
+      expect(await user.getIdToken(refresh: true), equals(kMockIdToken));
+>>>>>>> origin/master
       expect(
         log,
         <Matcher>[
-          isMethodCall('signInAnonymously',
+          isMethodCall('currentUser',
               arguments: <String, String>{'app': auth.app.name}),
           isMethodCall(
             'getIdToken',
@@ -142,6 +172,18 @@ void main() {
             'getIdToken',
             arguments: <String, dynamic>{'refresh': true, 'app': auth.app.name},
           ),
+        ],
+      );
+    });
+
+    test('signInAnonymously', () async {
+      final AuthResult result = await auth.signInAnonymously();
+      verifyAuthResult(result);
+      expect(
+        log,
+        <Matcher>[
+          isMethodCall('signInAnonymously',
+              arguments: <String, String>{'app': auth.app.name}),
         ],
       );
     });
@@ -186,10 +228,11 @@ void main() {
     });
 
     test('signInWithEmailAndLink', () async {
-      await auth.signInWithEmailAndLink(
+      final AuthResult result = await auth.signInWithEmailAndLink(
         email: 'test@example.com',
         link: '<Url with domain from your Firebase project>',
       );
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -203,11 +246,11 @@ void main() {
     });
 
     test('createUserWithEmailAndPassword', () async {
-      final FirebaseUser user = await auth.createUserWithEmailAndPassword(
+      final AuthResult result = await auth.createUserWithEmailAndPassword(
         email: kMockEmail,
         password: kMockPassword,
       );
-      verifyUser(user);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -247,9 +290,9 @@ void main() {
         email: 'test@example.com',
         link: '<Url with domain from your Firebase project>',
       );
-      FirebaseUser user = await auth.currentUser();
-      user = await user.linkWithCredential(credential);
-      verifyUser(user);
+      final FirebaseUser user = await auth.currentUser();
+      final AuthResult result = await user.linkWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -279,8 +322,8 @@ void main() {
         email: 'test@example.com',
         link: '<Url with domain from your Firebase project>',
       );
-      final FirebaseUser user = await auth.signInWithCredential(credential);
-      verifyUser(user);
+      final AuthResult result = await auth.signInWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -330,9 +373,9 @@ void main() {
         authToken: kMockIdToken,
         authTokenSecret: kMockAccessToken,
       );
-      FirebaseUser user = await auth.currentUser();
-      user = await user.linkWithCredential(credential);
-      verifyUser(user);
+      final FirebaseUser user = await auth.currentUser();
+      final AuthResult result = await user.linkWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -362,8 +405,8 @@ void main() {
         authToken: kMockIdToken,
         authTokenSecret: kMockAccessToken,
       );
-      final FirebaseUser user = await auth.signInWithCredential(credential);
-      verifyUser(user);
+      final AuthResult result = await auth.signInWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -386,9 +429,9 @@ void main() {
       final AuthCredential credential = GithubAuthProvider.getCredential(
         token: kMockGithubToken,
       );
-      FirebaseUser user = await auth.currentUser();
-      user = await user.linkWithCredential(credential);
-      verifyUser(user);
+      final FirebaseUser user = await auth.currentUser();
+      final AuthResult result = await user.linkWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -416,8 +459,8 @@ void main() {
       final AuthCredential credential = GithubAuthProvider.getCredential(
         token: kMockGithubToken,
       );
-      final FirebaseUser user = await auth.signInWithCredential(credential);
-      verifyUser(user);
+      final AuthResult result = await auth.signInWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -440,9 +483,9 @@ void main() {
         email: kMockEmail,
         password: kMockPassword,
       );
-      FirebaseUser user = await auth.currentUser();
-      user = await user.linkWithCredential(credential);
-      verifyUser(user);
+      final FirebaseUser user = await auth.currentUser();
+      final AuthResult result = await user.linkWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -472,8 +515,8 @@ void main() {
         idToken: kMockIdToken,
         accessToken: kMockAccessToken,
       );
-      final FirebaseUser user = await auth.signInWithCredential(credential);
-      verifyUser(user);
+      final AuthResult result = await auth.signInWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -497,7 +540,8 @@ void main() {
         verificationId: kMockVerificationId,
         smsCode: kMockSmsCode,
       );
-      await auth.signInWithCredential(credential);
+      final AuthResult result = await auth.signInWithCredential(credential);
+      verifyAuthResult(result);
       expect(log, <Matcher>[
         isMethodCall('signInWithCredential', arguments: <String, dynamic>{
           'app': auth.app.name,
@@ -536,7 +580,9 @@ void main() {
         email: kMockEmail,
         password: kMockPassword,
       );
-      await user.reauthenticateWithCredential(credential);
+      final AuthResult result =
+          await user.reauthenticateWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -561,7 +607,9 @@ void main() {
         idToken: kMockIdToken,
         accessToken: kMockAccessToken,
       );
-      await user.reauthenticateWithCredential(credential);
+      final AuthResult result =
+          await user.reauthenticateWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -586,7 +634,9 @@ void main() {
       final AuthCredential credential = FacebookAuthProvider.getCredential(
         accessToken: kMockAccessToken,
       );
-      await user.reauthenticateWithCredential(credential);
+      final AuthResult result =
+          await user.reauthenticateWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -611,7 +661,9 @@ void main() {
         authToken: kMockAuthToken,
         authTokenSecret: kMockAuthTokenSecret,
       );
-      await user.reauthenticateWithCredential(credential);
+      final AuthResult result =
+          await user.reauthenticateWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -636,7 +688,9 @@ void main() {
       final AuthCredential credential = GithubAuthProvider.getCredential(
         token: kMockGithubToken,
       );
-      await user.reauthenticateWithCredential(credential);
+      final AuthResult result =
+          await user.reauthenticateWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -659,9 +713,9 @@ void main() {
         idToken: kMockIdToken,
         accessToken: kMockAccessToken,
       );
-      FirebaseUser user = await auth.currentUser();
-      user = await user.linkWithCredential(credential);
-      verifyUser(user);
+      final FirebaseUser user = await auth.currentUser();
+      final AuthResult result = await user.linkWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -690,9 +744,9 @@ void main() {
       final AuthCredential credential = FacebookAuthProvider.getCredential(
         accessToken: kMockAccessToken,
       );
-      FirebaseUser user = await auth.currentUser();
-      user = await user.linkWithCredential(credential);
-      verifyUser(user);
+      final FirebaseUser user = await auth.currentUser();
+      final AuthResult result = await user.linkWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -720,8 +774,8 @@ void main() {
       final AuthCredential credential = FacebookAuthProvider.getCredential(
         accessToken: kMockAccessToken,
       );
-      final FirebaseUser user = await auth.signInWithCredential(credential);
-      verifyUser(user);
+      final AuthResult result = await auth.signInWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -744,9 +798,9 @@ void main() {
         authToken: kMockAuthToken,
         authTokenSecret: kMockAuthTokenSecret,
       );
-      FirebaseUser user = await auth.currentUser();
-      user = await user.linkWithCredential(credential);
-      verifyUser(user);
+      final FirebaseUser user = await auth.currentUser();
+      final AuthResult result = await user.linkWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -776,8 +830,8 @@ void main() {
         authToken: kMockAuthToken,
         authTokenSecret: kMockAuthTokenSecret,
       );
-      final FirebaseUser user = await auth.signInWithCredential(credential);
-      verifyUser(user);
+      final AuthResult result = await auth.signInWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -800,9 +854,9 @@ void main() {
       final AuthCredential credential = GithubAuthProvider.getCredential(
         token: kMockGithubToken,
       );
-      FirebaseUser user = await auth.currentUser();
-      user = await user.linkWithCredential(credential);
-      verifyUser(user);
+      final FirebaseUser user = await auth.currentUser();
+      final AuthResult result = await user.linkWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -830,8 +884,8 @@ void main() {
       final AuthCredential credential = GithubAuthProvider.getCredential(
         token: kMockGithubToken,
       );
-      final FirebaseUser user = await auth.signInWithCredential(credential);
-      verifyUser(user);
+      final AuthResult result = await auth.signInWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -854,9 +908,9 @@ void main() {
         email: kMockEmail,
         password: kMockPassword,
       );
-      FirebaseUser user = await auth.currentUser();
-      user = await user.linkWithCredential(credential);
-      verifyUser(user);
+      final FirebaseUser user = await auth.currentUser();
+      final AuthResult result = await user.linkWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -1150,9 +1204,9 @@ void main() {
     });
 
     test('signInWithCustomToken', () async {
-      final FirebaseUser user =
+      final AuthResult result =
           await auth.signInWithCustomToken(token: kMockCustomToken);
-      verifyUser(user);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -1191,7 +1245,7 @@ void main() {
       await Future<void>.delayed(const Duration(seconds: 0));
 
       await simulateEvent(null);
-      await simulateEvent(mockFirebaseUser());
+      await simulateEvent(kMockUser);
 
       final FirebaseUser user1 = await events.remove();
       expect(user1, isNull);
@@ -1239,6 +1293,7 @@ void main() {
   });
 }
 
+<<<<<<< HEAD
 Map<String, dynamic> mockFirebaseUser(
         {String providerId = kMockProviderId,
         String uid = kMockUid,
@@ -1268,6 +1323,8 @@ Map<String, dynamic> mockIdTokenResult() => <String, dynamic>{
       'claims': kMockIdTokenResultClaims,
     };
 
+=======
+>>>>>>> origin/master
 /// Queue whose remove operation is asynchronous, awaiting a corresponding add.
 class AsyncQueue<T> {
   Map<int, Completer<T>> _completers = <int, Completer<T>>{};